Fix int64-limit round-trip: exact parsing and single sign
Two defects near the int64 limits:

- ParseUnit accumulated into a float64. Unit multipliers reach 1000^6
  (1e18) and 1024^6 (2^60), both beyond the 53-bit float64 mantissa, so
  MaxInt64 rounded up on parse and int64() wrapped the result to
  MinInt64. Accumulate with math/big instead and truncate toward zero
  with strict overflow detection.

- ToString formatted each component from the signed value, so negatives
  emitted a '-' per unit (e.g. "-9EB-223PB-...") — a string ToString's
  own parser rejects. Format the unsigned magnitude (uint64, so MinInt64
  is representable) and prepend a single leading '-'.

MakeUnitMap now returns map[string]int64 so multipliers stay exact.

+++ si.go
@@ -13,15 +13,15 @@ const (
 	Exa     = Peta * 1000
 )
 
-func MakeUnitMap(suffix, shortSuffix string, scale int64) map[string]float64 {
-	res := map[string]float64{
+func MakeUnitMap(suffix, shortSuffix string, scale int64) map[string]int64 {
+	res := map[string]int64{
 		shortSuffix: 1,
 		// see below for "k" / "K"
-		"M" + suffix: float64(scale * scale),
-		"G" + suffix: float64(scale * scale * scale),
-		"T" + suffix: float64(scale * scale * scale * scale),
-		"P" + suffix: float64(scale * scale * scale * scale * scale),
-		"E" + suffix: float64(scale * scale * scale * scale * scale * scale),
+		"M" + suffix: scale * scale,
+		"G" + suffix: scale * scale * scale,
+		"T" + suffix: scale * scale * scale * scale,
+		"P" + suffix: scale * scale * scale * scale * scale,
+		"E" + suffix: scale * scale * scale * scale * scale * scale,
 	}
 
 	// Standard SI prefixes use lowercase "k" for kilo = 1000.
@@ -41,10 +41,10 @@ func MakeUnitMap(suffix, shortSuffix string, scale int64) map[string]float64 {
 	//     -- https://en.wikipedia.org/wiki/Binary_prefix#History
 	//     See also the extensive https://en.wikipedia.org/wiki/Timeline_of_binary_prefixes.
 	if scale == 1024 {
-		res["K"+suffix] = float64(scale)
+		res["K"+suffix] = scale
 	} else {
-		res["k"+suffix] = float64(scale)
-		res["K"+suffix] = float64(scale)
+		res["k"+suffix] = scale
+		res["K"+suffix] = scale
 	}
 	return res
 }
